+++ packages/core/src/internal/utils.js
@@ -80,23 +80,18 @@ export function makeIterator(next, thro = kThrow, name = 'iterator') {
   return iterator
 }
 
-/**
-  Print error in a useful way whether in a browser environment
-  (with expandable error stack traces), or in a node.js environment
-  (text-only log output)
- **/
-export function log(level, message, error = '') {
+export function logError(error) {
   /*eslint-disable no-console*/
-  if (typeof window === 'undefined') {
-    console.log(`redux-saga ${level}: ${message}\n${(error && error.stack) || error}`)
-  } else {
-    console[level](message, error)
+  console.error(error)
+
+  if (error && error.sagaStack) {
+    console.error(error.sagaStack)
   }
 }
 
 export function deprecate(fn, deprecationWarning) {
   return (...args) => {
-    if (process.env.NODE_ENV !== 'production') log('warn', deprecationWarning)
+    if (process.env.NODE_ENV !== 'production') console.warn(deprecationWarning)
     return fn(...args)
   }
 }